import AnyEvent::Gearman 0.03001 from CPAN
git-cpan-module: AnyEvent::Gearman
git-cpan-version: 0.03001
git-cpan-authorid: TYPESTER

+++ t/simple.t
@@ -56,10 +56,13 @@ if (!defined $child) {
 }
 elsif ($child == 0) {
     my $server = Gearman::Server->new( port => $port );
-    $server->start_worker('t/danga_worker.pl -s 127.0.0.1:'.$port);
+    $server->start_worker("$^X t/danga_worker.pl -s 127.0.0.1:$port");
     Danga::Socket->EventLoop;
 }
+else {
+    END { kill 9, $child if $child }
+}
+
+sleep 1;
 
 run_tests;
-
-kill 9, $child;
